refactor: standardize header file inclusion style
Changed the format of include directives throughout the repository. Previously, both angle brackets (<>) and double quotes ("") were used inconsistently. Now, angle brackets are used for standard library, platform-associated, and third-party library header files that are installed in standard locations. Double quotes are used for programmer-defined and project-specific header files. This change aligns with the best practice as discussed on StackOverflow: https://stackoverflow.com/questions/21593/what-is-the-difference-between-include-filename-and-include-filename